ecosystem/skills/elevenlabs/requirements.txt
PyPDF2>=3.0.0
python-docx>=1.0.0
